refactor(PLM-166): Make None check a keyword option
[Why]
Sometime we might need multiple checks in a checker such as STR, LENGTH, NONE
And we use the op.ANY to validate the result if its either a string or None, but the results of STR and LENGHT must be combined togehter.
In this case, a list which passes the LENGTH check may lead to a wrong validation result.

[How]
Make None a keyword, so that we can satisfy more cases.

[PLM-166]